requirements.txt
shazamio
sounddevice
requests
Pillow
scipy
pygame
